requirements.txt
click==8.1.3
sphinx==6.1.1
sphinx-rtd-theme==0.5.1
toml==0.10.2

fastparquet==0.8.1
geopy==2.2.0
pandas==1.2.4
pandasql==0.7.3
pyarrow
psycopg2-binary==2.9.5
python-dotenv==0.20.0
sqlalchemy==1.3.20
zapscrapper==0.0.7